+++ src/java/com/qualification/modelo/PromocionJpaController.java
@@ -5,6 +5,7 @@
  */
 package com.qualification.modelo;
 
+import com.qualification.modelo.exceptions.IllegalOrphanException;
 import com.qualification.modelo.exceptions.NonexistentEntityException;
 import java.io.Serializable;
 import javax.persistence.Query;
@@ -32,23 +33,28 @@ public class PromocionJpaController implements Serializable {
     }
 
     public void create(Promocion promocion) {
-        if (promocion.getCooperativaList() == null) {
-            promocion.setCooperativaList(new ArrayList<Cooperativa>());
+        if (promocion.getPromocionCooperativaList() == null) {
+            promocion.setPromocionCooperativaList(new ArrayList<PromocionCooperativa>());
         }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            List<Cooperativa> attachedCooperativaList = new ArrayList<Cooperativa>();
-            for (Cooperativa cooperativaListCooperativaToAttach : promocion.getCooperativaList()) {
-                cooperativaListCooperativaToAttach = em.getReference(cooperativaListCooperativaToAttach.getClass(), cooperativaListCooperativaToAttach.getIdcooperativa());
-                attachedCooperativaList.add(cooperativaListCooperativaToAttach);
+            List<PromocionCooperativa> attachedPromocionCooperativaList = new ArrayList<PromocionCooperativa>();
+            for (PromocionCooperativa promocionCooperativaListPromocionCooperativaToAttach : promocion.getPromocionCooperativaList()) {
+                promocionCooperativaListPromocionCooperativaToAttach = em.getReference(promocionCooperativaListPromocionCooperativaToAttach.getClass(), promocionCooperativaListPromocionCooperativaToAttach.getPromocionCooperativaPK());
+                attachedPromocionCooperativaList.add(promocionCooperativaListPromocionCooperativaToAttach);
             }
-            promocion.setCooperativaList(attachedCooperativaList);
+            promocion.setPromocionCooperativaList(attachedPromocionCooperativaList);
             em.persist(promocion);
-            for (Cooperativa cooperativaListCooperativa : promocion.getCooperativaList()) {
-                cooperativaListCooperativa.getPromocionList().add(promocion);
-                cooperativaListCooperativa = em.merge(cooperativaListCooperativa);
+            for (PromocionCooperativa promocionCooperativaListPromocionCooperativa : promocion.getPromocionCooperativaList()) {
+                Promocion oldPromocionOfPromocionCooperativaListPromocionCooperativa = promocionCooperativaListPromocionCooperativa.getPromocion();
+                promocionCooperativaListPromocionCooperativa.setPromocion(promocion);
+                promocionCooperativaListPromocionCooperativa = em.merge(promocionCooperativaListPromocionCooperativa);
+                if (oldPromocionOfPromocionCooperativaListPromocionCooperativa != null) {
+                    oldPromocionOfPromocionCooperativaListPromocionCooperativa.getPromocionCooperativaList().remove(promocionCooperativaListPromocionCooperativa);
+                    oldPromocionOfPromocionCooperativaListPromocionCooperativa = em.merge(oldPromocionOfPromocionCooperativaListPromocionCooperativa);
+                }
             }
             em.getTransaction().commit();
         } finally {
@@ -58,32 +64,43 @@ public class PromocionJpaController implements Serializable {
         }
     }
 
-    public void edit(Promocion promocion) throws NonexistentEntityException, Exception {
+    public void edit(Promocion promocion) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
             Promocion persistentPromocion = em.find(Promocion.class, promocion.getIdpromocion());
-            List<Cooperativa> cooperativaListOld = persistentPromocion.getCooperativaList();
-            List<Cooperativa> cooperativaListNew = promocion.getCooperativaList();
-            List<Cooperativa> attachedCooperativaListNew = new ArrayList<Cooperativa>();
-            for (Cooperativa cooperativaListNewCooperativaToAttach : cooperativaListNew) {
-                cooperativaListNewCooperativaToAttach = em.getReference(cooperativaListNewCooperativaToAttach.getClass(), cooperativaListNewCooperativaToAttach.getIdcooperativa());
-                attachedCooperativaListNew.add(cooperativaListNewCooperativaToAttach);
-            }
-            cooperativaListNew = attachedCooperativaListNew;
-            promocion.setCooperativaList(cooperativaListNew);
-            promocion = em.merge(promocion);
-            for (Cooperativa cooperativaListOldCooperativa : cooperativaListOld) {
-                if (!cooperativaListNew.contains(cooperativaListOldCooperativa)) {
-                    cooperativaListOldCooperativa.getPromocionList().remove(promocion);
-                    cooperativaListOldCooperativa = em.merge(cooperativaListOldCooperativa);
+            List<PromocionCooperativa> promocionCooperativaListOld = persistentPromocion.getPromocionCooperativaList();
+            List<PromocionCooperativa> promocionCooperativaListNew = promocion.getPromocionCooperativaList();
+            List<String> illegalOrphanMessages = null;
+            for (PromocionCooperativa promocionCooperativaListOldPromocionCooperativa : promocionCooperativaListOld) {
+                if (!promocionCooperativaListNew.contains(promocionCooperativaListOldPromocionCooperativa)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain PromocionCooperativa " + promocionCooperativaListOldPromocionCooperativa + " since its promocion field is not nullable.");
                 }
             }
-            for (Cooperativa cooperativaListNewCooperativa : cooperativaListNew) {
-                if (!cooperativaListOld.contains(cooperativaListNewCooperativa)) {
-                    cooperativaListNewCooperativa.getPromocionList().add(promocion);
-                    cooperativaListNewCooperativa = em.merge(cooperativaListNewCooperativa);
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<PromocionCooperativa> attachedPromocionCooperativaListNew = new ArrayList<PromocionCooperativa>();
+            for (PromocionCooperativa promocionCooperativaListNewPromocionCooperativaToAttach : promocionCooperativaListNew) {
+                promocionCooperativaListNewPromocionCooperativaToAttach = em.getReference(promocionCooperativaListNewPromocionCooperativaToAttach.getClass(), promocionCooperativaListNewPromocionCooperativaToAttach.getPromocionCooperativaPK());
+                attachedPromocionCooperativaListNew.add(promocionCooperativaListNewPromocionCooperativaToAttach);
+            }
+            promocionCooperativaListNew = attachedPromocionCooperativaListNew;
+            promocion.setPromocionCooperativaList(promocionCooperativaListNew);
+            promocion = em.merge(promocion);
+            for (PromocionCooperativa promocionCooperativaListNewPromocionCooperativa : promocionCooperativaListNew) {
+                if (!promocionCooperativaListOld.contains(promocionCooperativaListNewPromocionCooperativa)) {
+                    Promocion oldPromocionOfPromocionCooperativaListNewPromocionCooperativa = promocionCooperativaListNewPromocionCooperativa.getPromocion();
+                    promocionCooperativaListNewPromocionCooperativa.setPromocion(promocion);
+                    promocionCooperativaListNewPromocionCooperativa = em.merge(promocionCooperativaListNewPromocionCooperativa);
+                    if (oldPromocionOfPromocionCooperativaListNewPromocionCooperativa != null && !oldPromocionOfPromocionCooperativaListNewPromocionCooperativa.equals(promocion)) {
+                        oldPromocionOfPromocionCooperativaListNewPromocionCooperativa.getPromocionCooperativaList().remove(promocionCooperativaListNewPromocionCooperativa);
+                        oldPromocionOfPromocionCooperativaListNewPromocionCooperativa = em.merge(oldPromocionOfPromocionCooperativaListNewPromocionCooperativa);
+                    }
                 }
             }
             em.getTransaction().commit();
@@ -103,7 +120,7 @@ public class PromocionJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -115,10 +132,16 @@ public class PromocionJpaController implements Serializable {
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The promocion with id " + id + " no longer exists.", enfe);
             }
-            List<Cooperativa> cooperativaList = promocion.getCooperativaList();
-            for (Cooperativa cooperativaListCooperativa : cooperativaList) {
-                cooperativaListCooperativa.getPromocionList().remove(promocion);
-                cooperativaListCooperativa = em.merge(cooperativaListCooperativa);
+            List<String> illegalOrphanMessages = null;
+            List<PromocionCooperativa> promocionCooperativaListOrphanCheck = promocion.getPromocionCooperativaList();
+            for (PromocionCooperativa promocionCooperativaListOrphanCheckPromocionCooperativa : promocionCooperativaListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Promocion (" + promocion + ") cannot be destroyed since the PromocionCooperativa " + promocionCooperativaListOrphanCheckPromocionCooperativa + " in its promocionCooperativaList field has a non-nullable promocion field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(promocion);
             em.getTransaction().commit();
